Fully separate Constants.java into separate constants classes. Also begin feature-flagging scoring and intake
