+++ src/main/java/frc/robot/Robot.java
@@ -8,12 +8,9 @@ import edu.wpi.first.wpilibj.PowerDistribution;
 import edu.wpi.first.wpilibj.PowerDistribution.ModuleType;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.CommandScheduler;
-import frc.robot.constants.Constants;
-import org.littletonrobotics.junction.LogFileUtil;
 import org.littletonrobotics.junction.LoggedRobot;
 import org.littletonrobotics.junction.Logger;
 import org.littletonrobotics.junction.networktables.NT4Publisher;
-import org.littletonrobotics.junction.wpilog.WPILOGReader;
 import org.littletonrobotics.junction.wpilog.WPILOGWriter;
 
 public class Robot extends LoggedRobot {
@@ -32,24 +29,26 @@ public class Robot extends LoggedRobot {
             Logger.addDataReceiver(new WPILOGWriter()); // Log to a USB stick ("/U/logs")
             Logger.addDataReceiver(new NT4Publisher()); // Publish data to NetworkTables
             new PowerDistribution(1, ModuleType.kRev); // Enables power distribution logging
-        } else if (Constants.currentMode == Constants.Mode.SIM) {
+        } else {
             setUseTiming(false);
             Logger.addDataReceiver(new WPILOGWriter("logs/")); // This folder is gitignored
             Logger.addDataReceiver(new NT4Publisher());
-        } else {
-            setUseTiming(false); // Run as fast as possible
-            String logPath = LogFileUtil.findReplayLog(); // Pull the replay log from AdvantageScope
-            // (or prompt the user)
-            Logger.setReplaySource(new WPILOGReader(logPath)); // Read replay log
-            Logger.addDataReceiver(
-                    new WPILOGWriter(LogFileUtil.addPathSuffix(logPath, "_sim"))); // Save
-            // outputs
-            // to
-            // a
-            // new
-            // log
-        }
-        Logger.start();
+        } /*TODO: Fix replay mode!
+          else {
+              setUseTiming(false); // Run as fast as possible
+              String logPath = LogFileUtil.findReplayLog(); // Pull the replay log from AdvantageScope
+              // (or prompt the user)
+              Logger.setReplaySource(new WPILOGReader(logPath)); // Read replay log
+              Logger.addDataReceiver(
+                      new WPILOGWriter(LogFileUtil.addPathSuffix(logPath, "_sim"))); // Save
+              // outputs
+              // to
+              // a
+              // new
+              // log
+          }
+          Logger.start();
+          */
     }
 
     @Override
