+++ src/main/java/frc/robot/commands/ShootWithGamepad.java
@@ -2,7 +2,7 @@ package frc.robot.commands;
 
 import edu.wpi.first.wpilibj.DriverStation;
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.constants.Constants.AlignTarget;
+import frc.robot.constants.PhoenixDriveConstants.AlignTarget;
 import frc.robot.subsystems.scoring.ScoringSubsystem;
 import frc.robot.subsystems.scoring.ScoringSubsystem.ScoringAction;
 import java.util.function.BooleanSupplier;
