+++ src/main/java/frc/robot/constants/FeatureFlags.java
@@ -6,4 +6,8 @@ public final class FeatureFlags {
     public static final boolean simulateVision = true;
     public static final boolean runVision = true;
     public static final boolean runLocalizer = true;
+    public static final boolean simulateIntake = true;
+    public static final boolean runIntake = true;
+    public static final boolean simulateScoring = true;
+    public static final boolean runScoring = true;
 }
