+++ src/main/java/frc/robot/constants/SensorConstants.java
@@ -1,0 +1,7 @@
+package frc.robot.constants;
+
+public final class SensorConstants {
+    // TODO: Find real values
+    public static final int indexerSensorPort = 0;
+    public static final int uptakeSensorPort = 0;
+}
