+++ src/main/java/frc/robot/constants/SimConstants.java
@@ -1,0 +1,5 @@
+package frc.robot.constants;
+
+public final class SimConstants {
+    public static final double loopTime = 0.02;
+}
