+++ src/main/java/frc/robot/subsystems/drive/PhoenixDrive.java
@@ -25,8 +25,8 @@ import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.CommandScheduler;
 import edu.wpi.first.wpilibj2.command.Subsystem;
 import edu.wpi.first.wpilibj2.command.sysid.SysIdRoutine;
-import frc.robot.constants.Constants.AlignTarget;
 import frc.robot.constants.PhoenixDriveConstants;
+import frc.robot.constants.PhoenixDriveConstants.AlignTarget;
 import org.littletonrobotics.junction.Logger;
 
 public class PhoenixDrive extends SwerveDrivetrain implements Subsystem {
